requirements.txt
asgiref==3.5.0
autopep8==1.6.0
certifi==2022.5.18.1
cloudinary==1.29.0
Django==3.2
django-filter==21.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
Markdown==3.3.6
Pillow==9.0.1
pycodestyle==2.8.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
